backend/requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dialogflow==2.0.0
Flask==3.1.2
flask-cors==6.0.1
google-api-core==1.34.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.27.0
google-cloud-translate==3.21.1
google-cloud-vision==3.10.2
google-cloud-aiplatform==1.7.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.48.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
proto-plus==1.26.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
urllib3==2.5.0
Werkzeug==3.1.3
tqdm==4.67.1
typing-extensions==4.15.0
flask-socketio
eventlet
mediapipe
opencv-python-headless
firebase-admin
